requirements.txt
Django==3.2.10
djangorestframework==3.12.4
coverage==6.3.2
coverage==7.3.3